requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# LLM & RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.22
openai==1.10.0
tiktoken==0.5.2

# Database & Cache
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
supabase==2.0.0
pyodbc==5.0.1

# HTTP & API
requests==2.31.0
httpx==0.24.1

# Security & Rate Limiting
slowapi==0.1.9
python-jose[cryptography]==3.3.0

# Logging
structlog==24.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3
pandas==2.1.4

# Monitoring
prometheus-client==0.19.0
psutil==5.9.8
